Restructuring of the repository for better overview.
