+++ example_1/hardware/rrbot.hpp
