--- ros2_control_test_nodes/package.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>ros2_control_test_nodes</name>
-  <version>0.0.0</version>
-  <description>Demo nodes for showing and testing functionalities of the ros2_control framework.</description>
-
-  <maintainer email="denis@stogl.de">Denis Štogl</maintainer>
-
-  <license>Apache-2.0</license>
-
-  <depend>rclpy</depend>
-  <depend>std_msgs</depend>
-  <depend>trajectory_msgs</depend>
-
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
