--- ros2_control_test_nodes/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/ros2_control_test_nodes
-[install]
-install-scripts=$base/lib/ros2_control_test_nodes
